+++ CHANGES.rst
@@ -33,6 +33,10 @@ ChangeLog
 * Skip sets with unparseable mails because of incorrect encoding. Closes #47,
   #41 and #39.
 * Require the ``-s``/``--strategy`` CLI parameter to be set. Closes #44.
+* Add support for mboxes. Closes #48.
+* Remove requirement on ``-s``/``--strategy`` parameter, to let mails being
+  grouped into duplicate sets without any removal action, effectively acting as
+  a second-level dry-run.
 
 
 `2.0.1 (2016-11-28) <https://github.com/kdeldycke/maildir-deduplicate/compare/v2.0.0...v2.0.1>`_